test(picker-column-internal): migrate basic
